Copy from private repo
Copying code from private repo, changed a few things so might not be 100% working.

Remove MY_Log library for Sentry.io.

Moved tmp to be writable and updated code to reflect

Writable and Scritps are now in the document root with appropirate .htaccess files.

Added HOME_DIR as a local variable to allow for multilple hosting services.